switch between sm screen to <sm screens, loader to the icons before window object will work
