+++ components/hooks/useFocusTrap.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useEffect, RefObject } from 'react'
 
 const useFocusTrap = (
